requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
requests==2.31.0
httpx==0.26.0
python-dotenv==1.0.0
aiohttp==3.9.1
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
fuzzywuzzy
python-Levenshtein